refactor(resources): move secret CRUD routes
