+++ docs/context/architecture/multi-tenancy.md
@@ -12,6 +12,7 @@ sources:
   - src/treg/domain/identity/session.py
   - src/treg/routers/auth.py
   - src/treg/routers/orgs.py
+  - src/treg/routers/resources.py
   - src/treg/db.py
   - tests/test_router_dependencies.py
 related:
