+++ docs/context/interface/api.md
@@ -13,6 +13,7 @@ sources:
   - src/treg/routers/auth_helpers.py
   - src/treg/routers/catalog.py
   - src/treg/routers/orgs.py
+  - src/treg/routers/resources.py
   - src/treg/routers/signup_cookies.py
   - src/treg/routers/web.py
   - src/treg/domain/identity/access.py
@@ -116,7 +117,8 @@ the proven email, provisions the user, and commits before returning a framework-
 callbacks pop `_cli_states` only after that commit. The session-identity block translates
 `application.auth.current_identity` for `/auth/me`; `/auth/logout` remains an HTTP-only cookie action.
 `routers.orgs` translates signup, team-entry, invite-lifecycle, member-management, machine-identity,
-project, policy, settings, usage, and budget HTTP. `application.signup` owns the two creation
+project, policy, settings, usage, and budget HTTP; `routers.resources` translates secret CRUD HTTP.
+`application.signup` owns the two creation
 transactions and their promo, conversion, and referral sequence; `domain.governance.teams` owns team
 slug and membership creation plus the team-switcher read model.
