+++ src/treg/routers/resources.py
@@ -1,0 +1,157 @@
+"""Secret, tool, skill, and bundle HTTP routes."""
+
+import json
+
+from fastapi import APIRouter, Depends, HTTPException
+from pydantic import BaseModel
+from sqlalchemy.ext.asyncio import AsyncSession
+from sqlmodel import select
+
+from .. import crypto, sandbox as demo_sandbox
+from ..config import get_settings
+from ..db import get_session
+from ..domain.identity.access import Caller, _can_manage, _require_can_register, require_member
+from ..models import Bundle, Secret, Tool
+
+
+# The app alias preserves the moved handlers' original decorator text byte-for-byte.
+app = APIRouter()
+crud_router = app
+
+
+class SecretIn(BaseModel):
+    name: str
+    value: str
+    kind: str = "env"
+    bundle_id: int | None = None
+
+
+class SecretUpdate(BaseModel):
+    name: str | None = None
+    value: str | None = None
+    kind: str | None = None
+
+
+async def _visible_secret_ids(caller: Caller, db: AsyncSession) -> set[int] | None:
+    """The secret ids a tool-restricted member may SEE: the ones wired into their allowed tools
+    (HTTP bindings + cli.inject). None = unrestricted (owner / NULL tool_access) — show all. The
+    ACL isn't just a call gate: listings must not reveal credentials the member can't use."""
+    if caller.role == "owner" or caller.membership.tool_access is None:
+        return None
+    tools = (await db.execute(select(Tool).where(Tool.org_id == caller.org_id))).scalars().all()
+    ids: set[int] = set()
+    for t in tools:
+        if not _tool_usable(caller, t):
+            continue
+        ids |= {b.get("secret_id") for b in (t.bindings or []) if b.get("secret_id") is not None}
+        ids |= {e.get("secret_id") for e in ((t.cli or {}).get("inject") or []) if e.get("secret_id") is not None}
+    return ids
+
+
+@app.post("/secrets")
+async def create_secret(
+    body: SecretIn, caller: Caller = Depends(require_member), db: AsyncSession = Depends(get_session)
+) -> dict:
+    _require_can_register(caller)
+    await _enforce_sandbox_cap(caller, Secret, demo_sandbox.MAX_SECRETS, "secrets", db)
+    await _validate_bundle_id(body.bundle_id, caller.org_id, db)
+    secret = Secret(
+        org_id=caller.org_id, name=body.name, owner=caller.email, kind=body.kind,
+        value=crypto.encrypt(body.value), bundle_id=body.bundle_id,
+    )
+    db.add(secret)
+    await db.commit()
+    return _secret_view(secret)
+
+
+@app.get("/secrets")
+async def list_secrets(
+    caller: Caller = Depends(require_member), db: AsyncSession = Depends(get_session)
+) -> list[dict]:
+    rows = (await db.execute(select(Secret).where(Secret.org_id == caller.org_id))).scalars().all()
+    visible = await _visible_secret_ids(caller, db)
+    if visible is not None:  # tool-restricted member: only the keys wired into their allowed tools
+        rows = [s for s in rows if s.id in visible]
+    return [_secret_view(s) for s in rows]
+
+
+@app.patch("/secrets/{secret_id}")
+async def update_secret(
+    secret_id: int,
+    body: SecretUpdate,
+    caller: Caller = Depends(require_member),
+    db: AsyncSession = Depends(get_session),
+) -> dict:
+    secret = await db.get(Secret, secret_id)
+    if secret is None or secret.org_id != caller.org_id:
+        raise HTTPException(status_code=404, detail="secret not found")
+    if not _can_manage(caller, secret):
+        raise HTTPException(status_code=403, detail="only the creator or an admin can edit this secret")
+    _require_not_live_demo_secret(caller, secret)
+    fields = body.model_dump(exclude_unset=True)
+    for k in ("name", "value", "kind"):  # these map to NOT-NULL columns; explicit null is a 422, not a 500
+        if k in fields and fields[k] is None:
+            raise HTTPException(status_code=422, detail=f"{k} cannot be null")
+    # A kind change drives refresh + health + extraction shape; validate a JSON-kind actually has a
+    # JSON value (else the tool silently 502s later) and reset the now-meaningless health verdict.
+    if "kind" in fields and fields["kind"] != secret.kind:
+        if fields["kind"] in ("oauth", "secret_file"):
+            raw = fields["value"] if "value" in fields else crypto.decrypt(secret.value)
+            try:
+                json.loads(raw)
+            except (ValueError, TypeError):
+                raise HTTPException(status_code=422, detail=f"kind {fields['kind']!r} needs a JSON value")
+        secret.health_status, secret.health_detail, secret.health_checked_at = "unknown", "", None
+    if "value" in fields:
+        fields["value"] = crypto.encrypt(fields["value"])  # re-encrypt on rotate
+        # The value is exactly what health measures — a rotation invalidates the prior verdict.
+        secret.health_status, secret.health_detail, secret.health_checked_at = "unknown", "", None
+    for k, v in fields.items():
+        setattr(secret, k, v)
+    await db.commit()
+    return _secret_view(secret)
+
+
+@app.delete("/secrets/{secret_id}")
+async def delete_secret(
+    secret_id: int, caller: Caller = Depends(require_member), db: AsyncSession = Depends(get_session)
+) -> dict:
+    secret = await db.get(Secret, secret_id)
+    if secret is None or secret.org_id != caller.org_id:
+        raise HTTPException(status_code=404, detail="secret not found")
+    if not _can_manage(caller, secret):
+        raise HTTPException(status_code=403, detail="only the creator or an admin can delete this secret")
+    _require_not_live_demo_secret(caller, secret)
+    # bindings live in a JSON column — scan tools IN THIS ORG (registry-scale N is small).
+    tools = (await db.execute(select(Tool).where(Tool.org_id == caller.org_id))).scalars().all()
+    if any(b.get("secret_id") == secret_id for t in tools for b in t.bindings):
+        raise HTTPException(status_code=409, detail="secret is referenced by a tool binding")
+    # a secret used only by a local-run inject (not an HTTP binding) would otherwise be silently
+    # deletable, breaking `treg run` — guard those references too.
+    if any((e.get("secret_id") == secret_id) for t in tools for e in ((t.cli or {}).get("inject") or [])):
+        raise HTTPException(status_code=409, detail="secret is referenced by a tool's local-run (cli) profile")
+    await db.delete(secret)
+    await db.commit()
+    return {"deleted": secret_id}
+
+
+def _require_not_live_demo_secret(caller: Caller, secret: Secret) -> None:
+    """Companion guard for the seeded STRIPE_KEY the live tool is bound to."""
+    if (demo_sandbox.is_sandbox(caller.org) and get_settings().demo_stripe_key
+            and secret.name == "STRIPE_KEY"):
+        raise HTTPException(status_code=403, detail=(
+            "STRIPE_KEY powers the live stripe demo — add your own keys instead"))
+
+
+async def _validate_bundle_id(bundle_id: int | None, org_id: int, db: AsyncSession) -> None:
+    """A resource may only attach to a bundle in its OWN org — else it'd be counted by, rendered in,
+    and swept up by a foreign org's bundle view/delete (org-scoping leak)."""
+    if bundle_id is None:
+        return
+    bundle = await db.get(Bundle, bundle_id)
+    if bundle is None or bundle.org_id != org_id:
+        raise HTTPException(status_code=422, detail=f"bundle_id {bundle_id} not found in this org")
+
+
+def _secret_view(s: Secret) -> dict:
+    return {"id": s.id, "name": s.name, "kind": s.kind, "owner": s.owner, "bundle_id": s.bundle_id}
